Move Sentry notify task to overrides
